レベルに応じて検索条件が複雑化するように変更
